move node parser back to graph class
